// libs/web/constants/enums.ts
export enum ETextVariant {
  xs = 'xs' /* 12px */,
  sm = 'sm' /* 14px */,
  base = 'base' /* 16px */,
  lg = 'lg' /* 18px */,
  xl = 'xl' /* 20px */,
  xl_2 = 'xl_2' /* 24px */,
  xl_3 = 'xl_3' /* 30px */,
  xl_4 = 'xl_4' /* 36px */,
  xl_5 = 'xl_5' /* 48px */,
  xl_6 = 'xl_6' /* 60px */,
}

export enum ETextColor {
  black = 'black',
  primary = 'primary',
  secondary = 'secondary',
  error = 'error',
  gray = 'gray',
  white = 'white',
}

export enum EStatus {
  INACTIVE = 'Inactive',
  ACTIVE = 'Active',
  SUSPENDED = 'Suspended',
}

export enum ETooltipType {
  error = 'error',
  warning = 'warning',
}